+++ include/vec2.h
@@ -98,9 +98,9 @@ public:
     // comparators
     bool operator==(const Vec2& other) const
     {
-        return tdm::abs(entries[0] - other.entries[0]) < EPSILON_6 &&
-               tdm::abs(entries[1] - other.entries[1]) < EPSILON_6;
+        return fEqualE6(entries[0], other.entries[0]) && fEqualE6(entries[1], other.entries[1]);
     }
+
     bool operator!=(const Vec2& other) const { return !(*this == other); }
 
     // vector functions
@@ -112,11 +112,13 @@ public:
 
     float lengthSquared() const { return (entries[0] * entries[0] + entries[1] * entries[1]); }
 
+    bool isUnit() const { return fEqualE5(lengthSquared(), 1.0f); }
+
     Vec2 normalized() const
     {
         float len = length();
 
-        if (abs(len - 0.0f) < EPSILON_6)
+        if (fEqualE6(len, 0.0f))
             return *this;
 
         return (*this / len);