+++ include/vec4.h
@@ -117,10 +117,8 @@ public:
     // comparators
     bool operator==(const Vec4& other) const
     {
-        return tdm::abs(entries[0] - other.entries[0]) < EPSILON_6 &&
-               tdm::abs(entries[1] - other.entries[1]) < EPSILON_6 &&
-               tdm::abs(entries[2] - other.entries[2]) < EPSILON_6 &&
-               tdm::abs(entries[3] - other.entries[3]) < EPSILON_6;
+        return fEqualE6(entries[0], other.entries[0]) && fEqualE6(entries[1], other.entries[1]) &&
+               fEqualE6(entries[2], other.entries[2]) && fEqualE6(entries[3], other.entries[3]);
     }
 
     bool operator!=(const Vec4& other) const { return !(*this == other); }
@@ -143,11 +141,13 @@ public:
         return (entries[0] * entries[0] + entries[1] * entries[1] + entries[2] * entries[2] + entries[3] * entries[3]);
     }
 
+    bool isUnit() const { return fEqualE5(lengthSquared(), 1.0f); }
+
     Vec4 normalized() const
     {
         float len = length();
 
-        if (abs(len - 0.0f) < EPSILON_6)
+        if (fEqualE6(len, 0.0f))
             return *this;
 
         return (*this) / len;
@@ -166,7 +166,7 @@ public:
         {
             return (*this / entries[3]);
         }
-        return *this; // todo: maybe do something else? direction vector.
+        return *this;
     }
 
     // gettters